requirements.txt
facebook-sdk==2.0.0
peewee==2.10.1
psycopg2==2.6.2
schedule==0.4.2